@@ -12,25 +12,29 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^7.2.7",
+    "@angular/animations": "^7.2.13",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
+    "@angular/common": "^7.2.13",
+    "@angular/compiler": "~7.2.13",
+    "@angular/core": "~7.2.13",
     "@angular/fire": "^5.1.1",
     "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "~7.1.0",
+    "@angular/forms": "~7.2.13",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/platform-browser": "~7.2.13",
+    "@angular/platform-browser-dynamic": "~7.2.13",
+    "@angular/router": "~7.2.13",
     "@ngx-gallery/core": "^4.0.1",
     "@ngx-gallery/gallerize": "^4.0.1",
     "@ngx-gallery/lightbox": "^4.0.1",
+    "@progress/kendo-angular-buttons": "3 - 4",
+    "@progress/kendo-angular-conversational-ui": "^1.1.1",
+    "@progress/kendo-angular-l10n": "^1.1.0",
     "@types/chance": "^1.0.1",
     "angular-in-memory-web-api": "^0.8.0",
     "angular-instantsearch": "^2.2.1",
     "chance": "^1.0.18",
+    "chat-engine": "^0.9.21",
     "core-js": "^2.6.5",
     "faker": "^4.1.0",
     "firebase": "^5.8.5",
@@ -40,15 +44,17 @@
     "ngx-img-cropper": "^7.0.3",
     "ngx-textarea-autosize": "^2.0.0",
     "ngx-virtual-scroller": "^1.0.16",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.4.0",
+    "rxjs-compat": "^6.0.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.8.26",
+    "@progress/kendo-theme-default": "latest"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.4",
     "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "~7.1.0",
+    "@angular/compiler-cli": "^7.2.13",
+    "@angular/language-service": "~7.2.13",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -63,6 +69,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.2.4"
   }
 }
