@@ -1,7 +1,6 @@
 import {Component} from '@angular/core';
 import {AuthService} from '../services/auth.service';
 import {Theme} from 'ngx-auth-firebaseui';
-import {ChatService} from '../services/chat.service';
 
 @Component({
     selector: 'app-nav',
@@ -12,7 +11,6 @@ export class NavComponent {
 
     theme = Theme;
 
-    constructor(public auth: AuthService,
-                public chatService: ChatService) {
+    constructor(public auth: AuthService) {
     }
 }
