@@ -2,12 +2,14 @@ import {NgModule} from '@angular/core';
 import {MaterialModule} from './material.module';
 import {FormsModule} from '@angular/forms';
 import {CommonModule} from '@angular/common';
+import {FlexLayoutModule} from '@angular/flex-layout';
 
 @NgModule({
     exports: [
         FormsModule,
         CommonModule,
-        MaterialModule
+        MaterialModule,
+        FlexLayoutModule
     ]
 })
 export class SharedModule {
